Переписал третий экран под MVVM
