--- homework-3/homework-3/ThirdVCData.swift
@@ -1,14 +1,0 @@
-//
-//  ThirdVCData.swift
-//  homework-3
-//
-//  Created by user188734 on 5/12/21.
-//
-
-struct ThirdVCData {
-    let sport = "Хоккей"
-    let myHobbies = "Путешествия, спорт, настолки, программирование, книжки, сериальчики, английский, инвестиции в криптовалюты"
-    let newHobbie = "Малая авиация ✈️"
-    
-}
-
